Parameterise coefficients
This script loads in our PDB protein structures and uses the RMSD as the loss function to optimise the intersection, distance and contact coefficients. It also contains a correction to the quat2mat function. The intersection has been found to having very little influence over the scoring function, so there's also been some experimentation with altering the influence of non-zero values (can be commented out though).